Bug 2027564 - validate decoded image before converting to 8-bit.  a=diannaS DONTBUILD
Original Revision: https://phabricator.services.mozilla.com/D291850

Differential Revision: https://phabricator.services.mozilla.com/D293602

+++ dom/media/MediaData.h
@@ -492,11 +492,11 @@ class VideoData : public MediaData {
     ~QuantizableBuffer();
 
    private:
-    void AllocateRecyclableData(size_t aLength);
+    void AllocateRecyclableData(uint32_t aLength);
 
     RefPtr<layers::BufferRecycleBin> mRecycleBin;
     UniquePtr<uint8_t[]> m8bpcPlanes;
-    size_t mAllocatedLength;
+    uint32_t mAllocatedLength;
   };
 
   // Constructs a VideoData object. If aImage is nullptr, creates a new Image
